requirements.txt
Pillow == 10.1.0
matplotlib == 3.7.1
nltk == 3.5
numpy == 1.22.4
wordcloud == 1.9.2
